tools: fix incorrect version history order
This fixes an error in parseYAML(text), the version sorting
coudn't be right as we compared an arrify string
(ie. a = ["v18.11, v16.7.0"])with an array of strings
(ie. b = ["v18.07", "v16.7.0"]) in versionSort(a, b).

minVersion(a) couldn't find the minimum version with an arrify
string like a = ["v18.11, v16.7.0"].
That's why incorrect version history orders sometimes appeared.

Fixes: https://github.com/nodejs/node/issues/45670